pyproject.toml
[project]
name = "solar-payoff"
version = "0.1.0"
description = "Track solar panel ROI using Octopus Energy and Solis Cloud APIs"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
